requirements.txt
discord-py>=2.5.2
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
motor>=3.7.0
paramiko>=3.5.1
psycopg2-binary>=2.9.10
pymongo>=4.12.1
python-dotenv>=1.1.0
trafilatura>=2.0.0
psutil>=5.9.0
discord-py>=2.5.2
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
motor>=3.7.0
paramiko>=3.5.1
psutil>=5.9.0
psycopg2-binary>=2.9.10
pymongo>=4.12.1
python-dotenv>=1.1.0
trafilatura>=2.0.0